Backend/requirements.txt
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2
django-cors-headers==4.9.0
python-dotenv==1.2.1
Pillow==12.1.0
pytest==9.0.2
pytest-django==4.11.1
